Move request inside the ScoresContainer again
